refactor : ui 요소들 디렉토리 새로 생성
