--- app/src/main/java/woowacourse/shopping/cart/CartItem.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.cart
-
-import woowacourse.shopping.model.ProductUIModel
-
-data class CartItem(val product: ProductUIModel)
