--- app/src/main/java/woowacourse/shopping/shopping/viewHolder/ItemViewHolder.kt
@@ -1,6 +1,0 @@
-package woowacourse.shopping.shopping.viewHolder
-
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-sealed class ItemViewHolder(view: View) : RecyclerView.ViewHolder(view)
