+++ app/src/main/java/woowacourse/shopping/ui/cart/CartAdapter.kt
@@ -1,0 +1,29 @@
+package woowacourse.shopping.ui.cart
+
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.model.ProductUIModel
+import woowacourse.shopping.ui.cart.viewHolder.CartViewHolder
+
+class CartAdapter(
+    private val cartItems: List<CartItem>,
+    private val onItemClick: (ProductUIModel) -> Unit,
+    private val onItemRemove: (Int) -> Unit
+) : RecyclerView.Adapter<CartViewHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CartViewHolder {
+        return CartViewHolder.from(parent, onItemClick, onItemRemove)
+    }
+
+    override fun getItemCount(): Int {
+        return cartItems.size
+    }
+
+    override fun onBindViewHolder(holder: CartViewHolder, position: Int) {
+        holder.bind(cartItems[position])
+    }
+
+    override fun getItemViewType(position: Int): Int {
+        return position
+    }
+}
