+++ app/src/main/java/woowacourse/shopping/ui/cart/CartItem.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.ui.cart
+
+import woowacourse.shopping.model.ProductUIModel
+
+data class CartItem(val product: ProductUIModel)
