+++ app/src/main/java/woowacourse/shopping/ui/productdetail/contract/ProductDetailContract.kt
@@ -1,0 +1,15 @@
+package woowacourse.shopping.ui.productdetail.contract
+
+import woowacourse.shopping.model.ProductUIModel
+
+interface ProductDetailContract {
+    interface View {
+        fun setProductDetail(product: ProductUIModel)
+    }
+
+    interface Presenter {
+        fun setUpProductDetail()
+        fun addProductToCart()
+        fun addProductToRecent()
+    }
+}
