+++ app/src/main/java/woowacourse/shopping/ui/recentProduct/RecentProductItem.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.ui.recentProduct
+
+import woowacourse.shopping.model.ProductUIModel
+
+class RecentProductItem(val product: ProductUIModel)
