+++ app/src/main/res/layout/activity_product_detail.xml
@@ -13,7 +13,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".productdetail.ProductDetailActivity">
+        tools:context=".ui.productdetail.ProductDetailActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
