+++ app/src/main/res/layout/activity_shopping.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".shopping.ShoppingActivity">
+        tools:context=".ui.shopping.ShoppingActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
